Fixed bug when assigning a float value to an int member.

+++ Holoville/HOTween/Plugins/Core/ABSTweenPlugin.cs
@@ -29,6 +29,7 @@ using System.Reflection;
 using FastDynamicMemberAccessor;
 #endif
 using Holoville.HOTween.Core;
+using UnityEngine;
 
 namespace Holoville.HOTween.Plugins.Core
 {
@@ -513,12 +514,12 @@ namespace Holoville.HOTween.Plugins.Core
                 catch (InvalidCastException)
                 {
                     // This happens only if a float is being assigned to an int.
-                    propInfo.SetValue(tweenObj.target, (int)Math.Floor((double)p_value), null);
+                    propInfo.SetValue(tweenObj.target, (int)Math.Floor((double)(float)p_value), null);
                 }
                 catch (ArgumentException)
                 {
                     // This happens only on iOS if a float is being assigned to an int.
-                    propInfo.SetValue(tweenObj.target, (int)Math.Floor((double)p_value), null);
+                    propInfo.SetValue(tweenObj.target, (int)Math.Floor((double)(float)p_value), null);
                 }
             }
             else
@@ -530,12 +531,12 @@ namespace Holoville.HOTween.Plugins.Core
                 catch (InvalidCastException)
                 {
                     // This happens only if a float is being assigned to an int.
-                    fieldInfo.SetValue(tweenObj.target, (int)Math.Floor((double)p_value));
+                    fieldInfo.SetValue(tweenObj.target, (int)Math.Floor((double)(float)p_value));
                 }
                 catch (ArgumentException)
                 {
                     // This happens only on iOS if a float is being assigned to an int.
-                    fieldInfo.SetValue(tweenObj.target, (int)Math.Floor((double)p_value));
+                    fieldInfo.SetValue(tweenObj.target, (int)Math.Floor((double)(float)p_value));
                 }
             }
 #else
@@ -550,12 +551,12 @@ namespace Holoville.HOTween.Plugins.Core
                     catch (InvalidCastException)
                     {
                         // This happens only if a float is being assigned to an int.
-                        propInfo.SetValue(tweenObj.target, (int)Math.Floor((double)p_value), null);
+                        propInfo.SetValue(tweenObj.target, (int)Math.Floor((double)(float)p_value), null);
                     }
                     catch (ArgumentException)
                     {
                         // This happens only on iOS if a float is being assigned to an int.
-                        propInfo.SetValue(tweenObj.target, (int)Math.Floor((double)p_value), null);
+                        propInfo.SetValue(tweenObj.target, (int)Math.Floor((double)(float)p_value), null);
                     }
                 }
                 else
@@ -567,12 +568,12 @@ namespace Holoville.HOTween.Plugins.Core
                     catch (InvalidCastException)
                     {
                         // This happens only if a float is being assigned to an int.
-                        fieldInfo.SetValue(tweenObj.target, (int)Math.Floor((double)p_value));
+                        fieldInfo.SetValue(tweenObj.target, (int)Math.Floor((double)(float)p_value));
                     }
                     catch (ArgumentException)
                     {
                         // This happens only on iOS if a float is being assigned to an int.
-                        fieldInfo.SetValue(tweenObj.target, (int)Math.Floor((double)p_value));
+                        fieldInfo.SetValue(tweenObj.target, (int)Math.Floor((double)(float)p_value));
                     }
                 }
             }
@@ -585,12 +586,12 @@ namespace Holoville.HOTween.Plugins.Core
                 catch (InvalidCastException)
                 {
                     // This happens only if a float is being assigned to an int.
-                    valAccessor.Set(tweenObj.target, (int)Math.Floor((double)p_value)); // OPTIMIZE store if it's int prior to this, so valAccessor doesn't even have to run to catch the error?
+                    valAccessor.Set(tweenObj.target, (int)Math.Floor((double)(float)p_value)); // OPTIMIZE store if it's int prior to this, so valAccessor doesn't even have to run to catch the error?
                 }
                 catch (ArgumentException)
                 {
                     // This happens only on iOS if a float is being assigned to an int, but is also here just to be sure.
-                    valAccessor.Set(tweenObj.target, (int)Math.Floor((double)p_value));
+                    valAccessor.Set(tweenObj.target, (int)Math.Floor((double)(float)p_value));
                 }
             }
 #endif
